moved FOCVoteTracker & added KDoc comments
