+++ freedomOfComputing/src/main/java/nl/tudelft/trustchain/foc/MainActivityFOC.kt
@@ -37,7 +37,6 @@ import nl.tudelft.ipv8.android.IPv8Android
 import nl.tudelft.ipv8.keyvault.PrivateKey
 import nl.tudelft.trustchain.foc.community.FOCCommunity
 import nl.tudelft.trustchain.foc.community.FOCVote
-import nl.tudelft.trustchain.foc.community.FOCVoteTracker
 import nl.tudelft.trustchain.foc.community.signVote
 import nl.tudelft.trustchain.foc.databinding.ActivityMainFocBinding
 import nl.tudelft.trustchain.foc.util.ExtensionUtils.Companion.APK_DOT_EXTENSION
@@ -101,7 +100,6 @@ open class MainActivityFOC : AppCompatActivity() {
             }
 
             binding.torrentCount.text = getString(R.string.torrentCount, torrentAmount)
-            voteTracker.loadState(cacheDir.absolutePath + "/vote-tracker" + DATA_DOT_EXTENSION)
             copyDefaultApp()
             showAllFiles()
 
@@ -152,13 +150,21 @@ open class MainActivityFOC : AppCompatActivity() {
         progressVisible = !progressVisible
     }
 
+    /**
+     * This method is called upon starting the FreedomOfComputing app or coming back to it from
+     * another screen.
+     */
     override fun onResume() {
         super.onResume()
         resumeUISettings()
         appGossiper?.resume()
+        voteTracker.loadState(cacheDir.absolutePath + "/vote-tracker" + DATA_DOT_EXTENSION)
         focCommunity?.sendPullVotesMessage()
     }
 
+    /**
+     * This method is called upon closing the FreedomOfComputing app or going to another screen.
+     */
     override fun onPause() {
         super.onPause()
         appGossiper?.pause()
@@ -240,6 +246,9 @@ open class MainActivityFOC : AppCompatActivity() {
         Toast.makeText(applicationContext, s, Toast.LENGTH_SHORT).show()
     }
 
+    /**
+     * This method is used to create the blue button for an APK once it is successfully torrent.
+     */
     private fun createSuccessfulTorrentButton(uri: Uri) {
         val fileName = getFileName(uri)
         val torrentListView = binding.contentMainActivityFocLayout.torrentList
@@ -307,6 +316,10 @@ open class MainActivityFOC : AppCompatActivity() {
         }
     }
 
+    /**
+     * This method is used to update the vote count for an APK. It isn't called anywhere in this
+     * file, but it gets called from FOCCommunity upon receiving gossiped votes.
+     */
     fun updateVoteCounts(fileName: String) {
         val torrentListView = binding.contentMainActivityFocLayout.torrentList
         val row = torrentListView.findViewById<LinearLayout>(fileName.hashCode()) ?: return
@@ -321,6 +334,9 @@ open class MainActivityFOC : AppCompatActivity() {
         Log.i("vote-gossip", "Vote Count updated!")
     }
 
+    /**
+     * This method is used to create the grey button for an APK when torrent is still in progress.
+     */
     fun createUnsuccessfulTorrentButton(torrentName: String) {
         val torrentListView = binding.contentMainActivityFocLayout.torrentList
         val row = LinearLayout(this)
@@ -357,6 +373,9 @@ open class MainActivityFOC : AppCompatActivity() {
         button.isAllCaps = false
     }
 
+    /**
+     * This method created the popup modal when you press and hold on an APK.
+     */
     private fun createAlertDialog(fileName: String) {
         val builder: AlertDialog.Builder = AlertDialog.Builder(this)
         builder.setTitle(getString(R.string.createAlertDialogTitle))
@@ -367,6 +386,10 @@ open class MainActivityFOC : AppCompatActivity() {
         builder.show()
     }
 
+    /**
+     * Method that created the modal that shows up when you click the + sign in the bottom right
+     * hand corner of the FreedomOfComputing app.
+     */
     private fun createDownloadDialog() {
         val builder: AlertDialog.Builder = AlertDialog.Builder(this)
         builder.setTitle(getString(R.string.createDownloadDialogTitle))
@@ -395,6 +418,9 @@ open class MainActivityFOC : AppCompatActivity() {
         builder.show()
     }
 
+    /**
+     * This method is called from the alert dialog modal, when the user presses DELETE.
+     */
     private fun deleteApkFile(fileName: String) {
         val files = applicationContext.cacheDir.listFiles()
         if (files != null) {
